Add support for regular gamepad input
Some games cannot work with dual analog input. (eg. Castlevania -
Symphony of the Night).

I reuse the mouse device because RetroArch doesn't givce plugins the
ability to add "features". I choose specifically the mouse options
because it's the only device type that can be specified per input port.

The default is now "gamepad" because it's the most compatible option and it
is the default in mednafen [1].

[1] http://mednafen.sourceforge.net/documentation/09x/psx.html

Signed-off-by: Saggi Mizrahi <ficoos@gmail.com>