Update renderer.cpp
Trying to fix issue #126.

In the past, there was a bug in the GL renderer caused by the 'display_off' flag not being respected, which resulted in us getting some frames that were expected to be black but were instead filled with garbage or a combo of garbage and past frames.

Currently the Vulkan renderer seems to be suffering from the same bug. Based on the comments, the intention is that it is supposed to be done in scanout_to_texture(), but what I think I see is cleanups but not the insurance of a black screen.

So, in good hacking fashion, I tried to make the 'render_target' with the size of (0, 0), which segfaults. So I settled with 64u, which was already there and I assume is the smallest it can be before it segfaults.
And this seems to result in the desired black screen.

The patch needs to be vetted by someone who knows Vulkan.