fix splotchy tomb raider skin with bilinear/3point
Those -0.5 texel offsets were added to reduce the artifacting that happens with these filtering methods when they sample outside of the desired texture (both of them sample texels surrounding the current texel and then average their values together; this can lead to grabbing outlines or other colors that are in neighboring textures on the same texture sheet). This does not fix JINC2, which samples 2 texels out. This behavior leads me to guess that Tomb Raider may use a palette texture for the gouraud shading and these algos are jumping into the adjacent palettes when they sample the neighboring texels. /shrug