Fix issue #211
GPU is now a struct and a global variable in the stack, like
in upstream.
The GPU's vram is no longer 'flexible', it's still at the end
of the struct but it's a ptr type rather than a zero size array.

Changed some libretro calls so SET_GEOMETRY is used in favor of
SET_AV_INFO, as per the API's recommendation.

Style nits here and there as well.