- New more aggressive variant of the MakeQuad/FindLine Hack
- Stores modified vertices for second triangle before rendering the first (software renderer modifies vertex order)
- Added code to toggle between disabled/default/aggressive (no UI support though)

Required for games which draw lines but use UVs that form a triangle in texture space (Dark Forces and Duke Nukem).
These rely on PS1 sampling rules to follow the upper or left side UVs when drawing the line of pixels.
New version of the hack forces these UVs into a line, both for the original triangle and the secondary one needed to make a quad.

Retains the older more conservative method as it works fine for most games and produces fewer artifacts.