parallel-psx: Fix precision bug with resolve roundtrip.
This is a big problem when using paletted textures. It is very easy to
trigger scenarios where we roundtrip resolve unscaled -> scaled ->
unscaled.

We need to round in pack_abgr1555 because we can get an off-by-one
otherwise, especially when resolving large upscaling factors.

Also, remove mediump from the resolve shaders, since we need pretty
decent precision to ensure there is a 1:1 roundtrip.